move scripts to hack
