+++ .gitignore
@@ -1,1 +1,1 @@
-scripts/tools/bin
+hack/tools/bin
