+++ hack/capi.go
@@ -1,5 +1,5 @@
 // +build capi
 
-package scripts
+package hack
 
 import _ "sigs.k8s.io/cluster-api"
