+++ hack/generated.patch
