+++ hack/go.mod
@@ -1,0 +1,5 @@
+module github.com/giantswarm/apiextensions/hack
+
+go 1.14
+
+require sigs.k8s.io/cluster-api v0.2.10
