+++ hack/tools/controller-gen/go.mod
@@ -1,0 +1,5 @@
+module github.com/giantswarm/apiextensions/hack/tools/code-generator
+
+go 1.14
+
+require sigs.k8s.io/controller-tools v0.2.4
