+++ hack/tools/deepcopy-gen/go.mod
@@ -1,0 +1,5 @@
+module github.com/giantswarm/apiextensions/hack/tools/deepcopy-gen
+
+go 1.14
+
+require k8s.io/code-generator v0.16.6
