+++ hack/tools/esc/go.mod
@@ -1,4 +1,4 @@
-module github.com/giantswarm/apiextensions/scripts/tools/code-generator
+module github.com/giantswarm/apiextensions/hack/tools/code-generator
 
 go 1.14
