+++ hack/tools/kustomize/go.mod
@@ -1,0 +1,5 @@
+module github.com/giantswarm/apiextensions/hack/tools/code-generator
+
+go 1.14
+
+require sigs.k8s.io/kustomize/kustomize/v3 v3.5.4
