--- scripts/go.mod
@@ -1,5 +1,0 @@
-module github.com/giantswarm/apiextensions/scripts
-
-go 1.14
-
-require sigs.k8s.io/cluster-api v0.2.10
