--- scripts/tools/client-gen/go.mod
@@ -1,5 +1,0 @@
-module github.com/giantswarm/apiextensions/scripts/tools/client-gen
-
-go 1.14
-
-require k8s.io/code-generator v0.16.6
