--- scripts/tools/controller-gen/go.mod
@@ -1,5 +1,0 @@
-module github.com/giantswarm/apiextensions/scripts/tools/code-generator
-
-go 1.14
-
-require sigs.k8s.io/controller-tools v0.2.4
