--- scripts/tools/kustomize/go.mod
@@ -1,5 +1,0 @@
-module github.com/giantswarm/apiextensions/scripts/tools/code-generator
-
-go 1.14
-
-require sigs.k8s.io/kustomize/kustomize/v3 v3.5.4
